Remove a couple of boogers.
